+++ app/(auth)/signup/page.tsx
@@ -4,7 +4,7 @@ import Form from "@/app/_components/Form";
 import useCookie from "@/app/_hooks/useCookie";
 import API from "@/app/_api/index";
 import { useRouter } from "next/navigation";
-import { useMutation } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useCallback } from "react";
 import useAuthStore from "@/app/_store/useAuthStore";
 import Oauth from "@/app/(auth)/_components/Oauth";
@@ -16,6 +16,7 @@ export default function SignupPage() {
 	const [, setAccessToken] = useCookie<string>("accessToken");
 	const [, setRefreshToken] = useCookie<string>("refreshToken");
 	const setUser = useAuthStore((state) => state.setUser);
+	const queryClient = useQueryClient();
 
 	const signupMutation = useMutation({
 		mutationFn: async (ctx: FormContext) => {
@@ -29,6 +30,8 @@ export default function SignupPage() {
 			return API["{teamId}/auth/signUp"].POST({}, payload);
 		},
 		onSuccess: (response) => {
+			queryClient.invalidateQueries({ queryKey: ["user"] });
+
 			setUser({
 				id: response.user.id,
 				email: response.user.email || "",
